+++ src/aui-widget/js/aui-widget-transition.js
@@ -59,8 +59,7 @@ WidgetTransition.ATTRS = {
         value: {
             show: 0,
             hide: 0
-        },
-        writeOnce: true
+        }
     },
 
     /**